Move `useIsGeckoEnabled` hook out of `RegionSelect`
