+++ packages/manager/src/components/RegionSelect/RegionSelect.test.tsx
@@ -16,8 +16,8 @@ describe('RegionSelect', () => {
     currentCapability: 'Linodes',
     disabled: false,
     errorText: '',
-    flags: {},
     helperText: '',
+    isGeckoLAEnabled: false,
     label: '',
     onChange: vi.fn(),
     regions,
