+++ packages/manager/src/components/RegionSelect/RegionSelect.types.ts
@@ -6,8 +6,6 @@ import type {
 } from '@linode/api-v4';
 import type { DisableItemOption, EnhancedAutocompleteProps } from '@linode/ui';
 import type React from 'react';
-// @todo: modularization - Update type FlagSet to import from `@linode/shared` package once available.
-import type { FlagSet } from 'src/featureFlags';
 
 export type RegionFilterValue =
   | 'distributed-AF'
@@ -41,10 +39,6 @@ export interface RegionSelectProps<
    */
   disabledRegions?: Record<string, DisableItemOption>;
   /**
-   * Feature Flags
-   */
-  flags: FlagSet;
-  /**
    * Used to override filtering done by the `currentCapability` prop
    * @todo Remove this after Object Storage Gen2.
    */
@@ -55,6 +49,10 @@ export interface RegionSelectProps<
    * @default false
    */
   ignoreAccountAvailability?: boolean;
+  /**
+   * `isGeckoLAEnabled` flag from `useIsGeckoEnabled` hook
+   */
+  isGeckoLAEnabled: boolean;
   label?: string;
   regionFilter?: RegionFilterValue;
   /**
@@ -84,10 +82,6 @@ export interface RegionMultiSelectProps
   currentCapability: Capabilities | undefined;
   disabledRegions?: Record<string, DisableItemOption>;
   /**
-   * Feature Flags
-   */
-  flags: FlagSet;
-  /**
    * Used to override filtering done by the `currentCapability` prop
    * @todo Remove this after Object Storage Gen2.
    */
@@ -99,6 +93,10 @@ export interface RegionMultiSelectProps
    */
   ignoreAccountAvailability?: boolean;
   isClearable?: boolean;
+  /**
+   * `isGeckoLAEnabled` flag from `useIsGeckoEnabled` hook
+   */
+  isGeckoLAEnabled: boolean;
   label?: string;
   onChange: (ids: string[]) => void;
   regions: Region[];
