+++ packages/manager/src/features/CloudPulse/Alerts/CreateAlert/GeneralInformation/RegionSelect.tsx
@@ -1,4 +1,5 @@
 import { useRegionsQuery } from '@linode/queries';
+import { useIsGeckoEnabled } from '@linode/shared';
 import * as React from 'react';
 import { Controller, useFormContext } from 'react-hook-form';
 
@@ -20,6 +21,10 @@ export const CloudPulseRegionSelect = (props: CloudViewRegionSelectProps) => {
   const { data: regions, isError, isLoading } = useRegionsQuery();
   const { control } = useFormContext<CreateAlertDefinitionForm>();
   const flags = useFlags();
+  const { isGeckoLAEnabled } = useIsGeckoEnabled(
+    flags.gecko2?.enabled,
+    flags.gecko2?.la
+  );
 
   return (
     <Controller
@@ -33,8 +38,8 @@ export const CloudPulseRegionSelect = (props: CloudViewRegionSelectProps) => {
             field.onChange(value?.id);
           }}
           currentCapability={undefined}
-          flags={flags}
           fullWidth
+          isGeckoLAEnabled={isGeckoLAEnabled}
           label="Region"
           loading={isLoading}
           placeholder="Select a Region"
