+++ packages/manager/src/features/Images/ImagesCreate/ImageUpload.tsx
@@ -5,6 +5,7 @@ import {
   useProfile,
   useRegionsQuery,
 } from '@linode/queries';
+import { useIsGeckoEnabled } from '@linode/shared';
 import {
   ActionsPanel,
   Box,
@@ -59,6 +60,10 @@ export const ImageUpload = () => {
   const dispatch = useDispatch<Dispatch>();
   const hasPendingUpload = usePendingUpload();
   const flags = useFlags();
+  const { isGeckoLAEnabled } = useIsGeckoEnabled(
+    flags.gecko2?.enabled,
+    flags.gecko2?.la
+  );
 
   const [uploadProgress, setUploadProgress] = useState<AxiosProgressEvent>();
   const cancelRef = React.useRef<(() => void) | null>(null);
@@ -267,8 +272,8 @@ export const ImageUpload = () => {
                   currentCapability="Object Storage" // Images use Object Storage as their storage backend
                   disableClearable
                   errorText={fieldState.error?.message}
-                  flags={flags}
                   ignoreAccountAvailability
+                  isGeckoLAEnabled={isGeckoLAEnabled}
                   label="Region"
                   onChange={(e, region) => field.onChange(region.id)}
                   regionFilter="core" // Images service will not be supported for Gecko Beta
