+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/NetworkingSummaryPanel/NetworkingSummaryPanel.tsx
@@ -1,10 +1,10 @@
 import { useLinodeQuery } from '@linode/queries';
+import { useIsGeckoEnabled } from '@linode/shared';
 import { Paper } from '@linode/ui';
 import Grid from '@mui/material/Grid2';
 import { styled, useTheme } from '@mui/material/styles';
 import * as React from 'react';
 
-import { useIsGeckoEnabled } from 'src/components/RegionSelect/RegionSelect.utils';
 import { useFlags } from 'src/hooks/useFlags';
 
 import { DNSResolvers } from './DNSResolvers';
@@ -20,7 +20,10 @@ export const LinodeNetworkingSummaryPanel = React.memo((props: Props) => {
   const theme = useTheme();
   // @todo maybe move this query closer to the consuming component
   const { data: linode } = useLinodeQuery(props.linodeId);
-  const { isGeckoLAEnabled } = useIsGeckoEnabled(flags);
+  const { isGeckoLAEnabled } = useIsGeckoEnabled(
+    flags.gecko2?.enabled,
+    flags.gecko2?.la
+  );
 
   if (!linode) {
     return null;
