+++ packages/manager/src/features/ObjectStorage/BucketLanding/BucketRegions.tsx
@@ -1,3 +1,4 @@
+import { useIsGeckoEnabled } from '@linode/shared';
 import * as React from 'react';
 
 import { RegionSelect } from 'src/components/RegionSelect/RegionSelect';
@@ -27,6 +28,10 @@ export const BucketRegions = (props: Props) => {
   const { isObjectStorageGen2Enabled } = useIsObjectStorageGen2Enabled();
 
   const flags = useFlags();
+  const { isGeckoLAEnabled } = useIsGeckoEnabled(
+    flags.gecko2?.enabled,
+    flags.gecko2?.la
+  );
 
   // Error could be: 1. General Regions error, 2. Field error, 3. Nothing
   const errorText = error || allRegionsError?.[0]?.reason;
@@ -40,7 +45,7 @@ export const BucketRegions = (props: Props) => {
       disableClearable
       disabled={disabled}
       errorText={errorText}
-      flags={flags}
+      isGeckoLAEnabled={isGeckoLAEnabled}
       label="Region"
       onBlur={onBlur}
       onChange={(e, region) => onChange(region.id)}
