+++ packages/manager/src/features/VPCs/VPCCreate/FormComponents/VPCTopSectionContent.tsx
@@ -1,3 +1,4 @@
+import { useIsGeckoEnabled } from '@linode/shared';
 import { TextField } from '@linode/ui';
 import { getQueryParamsFromQueryString } from '@linode/utilities';
 import * as React from 'react';
@@ -12,8 +13,8 @@ import { sendLinodeCreateFormInputEvent } from 'src/utilities/analytics/formEven
 import { VPC_CREATE_FORM_VPC_HELPER_TEXT } from '../../constants';
 import { StyledBodyTypography } from './VPCCreateForm.styles';
 
-import type { CreateVPCPayload } from '@linode/api-v4';
 import type { Region } from '@linode/api-v4';
+import type { CreateVPCPayload } from '@linode/api-v4';
 import type { LinodeCreateType } from '@linode/utilities';
 import type { LinodeCreateQueryParams } from 'src/features/Linodes/types';
 
@@ -27,6 +28,10 @@ export const VPCTopSectionContent = (props: Props) => {
   const { disabled, isDrawer, regions } = props;
   const location = useLocation();
   const flags = useFlags();
+  const { isGeckoLAEnabled } = useIsGeckoEnabled(
+    flags.gecko2?.enabled,
+    flags.gecko2?.la
+  );
   const isFromLinodeCreate = location.pathname.includes('/linodes/create');
   const queryParams = getQueryParamsFromQueryString<LinodeCreateQueryParams>(
     location.search
@@ -61,7 +66,7 @@ export const VPCTopSectionContent = (props: Props) => {
             currentCapability="VPCs"
             disabled={isDrawer ? true : disabled}
             errorText={fieldState.error?.message}
-            flags={flags}
+            isGeckoLAEnabled={isGeckoLAEnabled}
             onBlur={field.onBlur}
             onChange={(_, region) => field.onChange(region?.id ?? '')}
             regions={regions}
