+++ packages/shared/src/index.ts
@@ -1,1 +1,3 @@
-export * from './LinodeSelect';
+export * from './components';
+
+export * from './hooks';
